tests/frontend/selenium/requirements.txt
pytest
requests
psycopg[binary]
python-dotenv
selenium==4.34.2